refactor: renomeia arquivos e funções de diabetes
