+++ src/common/formatters/cpf.ts
@@ -1,0 +1,1 @@
+export const cpfFormatter = (value: string) : string => value.replace(/(\d{3})(\d{3})(\d{3})(\d{2})/, "$1.$2.$3-$4");
