+++ src/features/acf/modules/AcfDashboardPage/modules/PanelSelector/modules/dashboards/modules/table/modules/diabetes/modules/columns/columns.formatter.tsx
@@ -1,8 +1,7 @@
+import { RenderDateTagCell } from "@/helpers/lista-nominal/renderCell";
 import { parseDate, formatDate } from "@/common/time";
+import { cpfFormatter } from "@/common/formatters/cpf";
 
-export const cpfFormatter = (value: string) : string => value.replace(/(\d{3})(\d{3})(\d{3})(\d{2})/, "$1.$2.$3-$4");
-
-//TODO: Mover para arquivo certo
 export const CpfOrBirthdayFormatter = ({ value }: { value: string | null }) => {
     if (!value) return <div data-testid="empty-return">{value}</div>;
     if (value.length === 11)
@@ -15,4 +14,12 @@ export const CpfOrBirthdayFormatter = ({ value }: { value: string | null }) => {
         );
     return <div data-testid="fallback-return">{value}</div>;
 };
+
+export const DateRenderCell = ({ value }: { value: string | null })=> {
+    if (!value)
+        return <div data-testid="tag">
+            <RenderDateTagCell/>
+        </div>;
+    return <div>{formatDate(new Date(value.slice(0, 10)))}</div>;
+}
